requirements.txt
ipython
matplotlib
pytorch
tensorflow-gpu
baidu-aip
pyaudio
bs4
feather-format==0.4.0
pandas
flask